general cleanup of variable typing and code layout, added a sniffer to track if the sleep breaks (somehow) and added input examples

+++ action_list.cc
@@ -116,7 +116,7 @@ const actions actionList[] =
 	}
 };//actionList
 
-const std::map<std::array<int, 2>, actions> buttonMapping =
+const std::map<std::array<short, 2>, actions> buttonMapping =
 {
 	//idle
 	{{-1, -1}, actionList[0]},
